# 20% npm install speedup: cache extras by package.json string