@iarna 2017-06-26
This was merged into v2.0.7 as 839b56e. Thank you so much! =)